refactored controller and express adapter
