+++ server/src/index.js
@@ -2,68 +2,19 @@ const express = require('express');
 const app = express();
 const dotenv = require('dotenv');
 const cors = require('cors');
-dotenv.config()
+dotenv.config();
+
+const callback = require('./support/ExpressCallback');
+const { index, search, show } = require('./controllers/moviesController')
 
 const corsOptions = {
     exposedHeaders: 'X-Page, X-Total-Pages, X-Total'
 }
 app.use(cors(corsOptions));
 
-const { index, search, show } = require('./controllers/moviesController')
-
-app.get('/movies', function (controller) {
-    return (req, res, next) => {
-        const httpRequest = {
-            body: req.body,
-            query: req.query,
-            params: req.params,
-            headers: req.headers
-        }
-
-        controller(httpRequest).then(httpResponse => {
-            if (httpResponse.headers) {
-                res.set(httpResponse.headers);
-            }
-            res.status(httpResponse.statusCode).send(httpResponse.body);
-        }).catch(e => next(e));
-    }
-}(index));
-
-app.get('/movies/search', function (controller) {
-    return (req, res, next) => {
-        const httpRequest = {
-            body: req.body,
-            query: req.query,
-            params: req.params,
-            headers: req.headers
-        }
-
-        controller(httpRequest).then(httpResponse => {
-            if (httpResponse.headers) {
-                res.set(httpResponse.headers);
-            }
-            res.status(httpResponse.statusCode).send(httpResponse.body);
-        }).catch(e => next(e));
-    }
-}(search));
-
-app.get('/movies/:movie_id', function (controller) {
-    return (req, res, next) => {
-        const httpRequest = {
-            body: req.body,
-            query: req.query,
-            params: req.params,
-            headers: req.headers
-        }
-
-        controller(httpRequest).then(httpResponse => {
-            if (httpResponse.headers) {
-                res.set(httpResponse.headers);
-            }
-            res.status(httpResponse.statusCode).send(httpResponse.body);
-        }).catch(e => next(e));
-    }
-}(show));
+app.get('/movies', callback(index));
+app.get('/movies/search', callback(search));
+app.get('/movies/:movie_id', callback(show));
 
 app.use(function (err, req, res, next) {
     console.error(err.stack);
